+++ src/lib.rs
@@ -24,7 +24,6 @@
 //! };
 //! use bls12_381::Bls12;
 //! use ff::PrimeField;
-//! use rand::rngs::OsRng;
 //! use sha2::{Digest, Sha256};
 //!
 //! /// Our own SHA-256d gadget. Input and output are in little-endian bit order.
@@ -95,11 +94,13 @@
 //!     }
 //! }
 //!
+//! let mut rng = rand::rng();
+//!
 //! // Create parameters for our circuit. In a production deployment these would
 //! // be generated securely using a multiparty computation.
 //! let params = {
 //!     let c = MyCircuit { preimage: None };
-//!     groth16::generate_random_parameters::<Bls12, _, _>(c, &mut OsRng).unwrap()
+//!     groth16::generate_random_parameters::<Bls12, _, _>(c, &mut rng).unwrap()
 //! };
 //!
 //! // Prepare the verification key (for proof verification).
@@ -115,7 +116,7 @@
 //! };
 //!
 //! // Create a Groth16 proof with our parameters.
-//! let proof = groth16::create_random_proof(c, &params, &mut OsRng).unwrap();
+//! let proof = groth16::create_random_proof(c, &params, &mut rng).unwrap();
 //!
 //! // Pack the hash as inputs for proof verification.
 //! let hash_bits = multipack::bytes_to_bits_le(&hash);